+++ core/divelist.h
@@ -14,7 +14,6 @@ struct dive;
 extern void update_cylinder_related_info(struct dive *);
 extern void mark_divelist_changed(bool);
 extern int unsaved_changes(void);
-extern void remove_autogen_trips(void);
 extern int init_decompression(struct deco_state *ds, struct dive *dive);
 
 /* divelist core logic functions */
@@ -37,15 +36,13 @@ extern dive_trip_t *create_and_hookup_trip_from_dive(struct dive *dive);
 extern dive_trip_t *get_dives_to_autogroup(int start, int *from, int *to, bool *allocated);
 extern dive_trip_t *get_trip_for_new_dive(struct dive *new_dive, bool *allocated);
 extern void autogroup_dives(void);
-extern struct dive *merge_two_dives(struct dive *a, struct dive *b);
 extern bool consecutive_selected();
 extern void select_dive(struct dive *dive);
 extern void deselect_dive(struct dive *dive);
 extern void select_dives_in_trip(struct dive_trip *trip);
 extern void deselect_dives_in_trip(struct dive_trip *trip);
 extern void filter_dive(struct dive *d, bool shown);
-extern void combine_trips(struct dive_trip *trip_a, struct dive_trip *trip_b);
-extern dive_trip_t *combine_trips_create(struct dive_trip *trip_a, struct dive_trip *trip_b);
+extern dive_trip_t *combine_trips(struct dive_trip *trip_a, struct dive_trip *trip_b);
 extern struct dive *first_selected_dive();
 extern struct dive *last_selected_dive();
 extern bool is_trip_before_after(const struct dive *dive, bool before);