+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+- Core: shift dive time in correct direction [#1893]
+- Include average max depth in statistics
+- Fix bug in cloud save after removing dives from a trip
 - Dive: Perform more accurate OTU calculations, and include
   OTU calculations for rebreather dives [#1851 & #1865].
+- Mobile: UI for copy-paste
 - Mobile: add initial copy-paste support
 - Desktop: translate trip date
 ---